refactor(hir-ty): extract internal type model
